Enable network logs in release builds
