+++ datasource/src/debug/kotlin/gq/kirmanak/mealient/DebugModule.kt
@@ -10,25 +10,12 @@ import dagger.hilt.InstallIn
 import dagger.hilt.android.qualifiers.ApplicationContext
 import dagger.hilt.components.SingletonComponent
 import dagger.multibindings.IntoSet
-import gq.kirmanak.mealient.datasource.BuildConfig
-import gq.kirmanak.mealient.logging.Logger
 import okhttp3.Interceptor
-import okhttp3.logging.HttpLoggingInterceptor
 
 @Module
 @InstallIn(SingletonComponent::class)
 internal object DebugModule {
 
-    @Provides
-    @IntoSet
-    fun provideLoggingInterceptor(logger: Logger): Interceptor {
-        val interceptor = HttpLoggingInterceptor { message -> logger.v(tag = "OkHttp") { message } }
-        interceptor.level = when {
-            BuildConfig.LOG_NETWORK -> HttpLoggingInterceptor.Level.BODY
-            else -> HttpLoggingInterceptor.Level.BASIC
-        }
-        return interceptor
-    }
 
     @Provides
     @IntoSet
