+++ datasource/src/main/kotlin/gq/kirmanak/mealient/datasource/DataSourceModule.kt
@@ -5,13 +5,17 @@ import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
 import dagger.hilt.components.SingletonComponent
+import dagger.multibindings.IntoSet
 import gq.kirmanak.mealient.datasource.impl.MealieDataSourceImpl
 import gq.kirmanak.mealient.datasource.impl.MealieServiceKtor
 import gq.kirmanak.mealient.datasource.impl.NetworkRequestWrapperImpl
 import gq.kirmanak.mealient.datasource.impl.OkHttpBuilderImpl
 import gq.kirmanak.mealient.datasource.impl.TrustedCertificatesStoreImpl
+import gq.kirmanak.mealient.logging.Logger
 import kotlinx.serialization.json.Json
+import okhttp3.Interceptor
 import okhttp3.OkHttpClient
+import okhttp3.logging.HttpLoggingInterceptor
 import javax.inject.Singleton
 
 @Module
@@ -33,6 +37,17 @@ internal interface DataSourceModule {
         fun provideOkHttp(okHttpBuilder: OkHttpBuilderImpl): OkHttpClient =
             okHttpBuilder.buildOkHttp()
 
+        @Provides
+        @IntoSet
+        fun provideLoggingInterceptor(logger: Logger): Interceptor {
+            val interceptor =
+                HttpLoggingInterceptor { message -> logger.v(tag = "OkHttp") { message } }
+            interceptor.level = when {
+                BuildConfig.LOG_NETWORK -> HttpLoggingInterceptor.Level.BODY
+                else -> HttpLoggingInterceptor.Level.BASIC
+            }
+            return interceptor
+        }
     }
 
     @Binds
